requirements.txt
allure-pytest==2.14.3
pytest==8.4.1
pytest-playwright==0.7.0
playwright==1.53.0
pytest-rerunfailures==15.1
pydantic==2.11.7
email-validator==2.2.0
pydantic-settings==2.10.1
pytest-xdist==3.8.0
ui-coverage-tool==0.31.0